Rental site with uer login and control panel to add and manage ads.
t’s a laravel based site. All of its backend coding has been done by laravel and database system is built by Eloquent Orm. This is a house renting site. People can add, edit and delete advertisements, there is a control panel/ dashboard for doing these operations. The front end of the site is based on a template but all of its data are parsed from database. As this site is still in work it needs user registration, ad details page, and ad to favorite plugin. As my previous training included jquery validation, I will validate the forms with jquery and php.